concerttickets-kommagetallen

+++ startcode/b_concerttickets.py
@@ -11,6 +11,7 @@
 # Voorbeelduitvoer:
 # De totale prijs bedraagt 31.65 euro.
 print('hallo, welkom')
-prijs = int(input("Wat is je de prijs van je tickets in euro's? "))
+
+prijs = float(input("Wat is je de prijs van je tickets in euro's? "))
 Aantal = int(input("Hoeveel tickets wil je kopen? "))
 print("je moet ",   prijs * Aantal , "euro betalen")
